# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Proyecto Banco WB",
        "description": "Conocer funcionamiento en peticiones http a la Api.",
        "contact": {
            "email": "contacto@bancowb.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/cuentas/listar": {
            "get": {
                "tags": [
                    "Projects"
                ],
                "summary": "Get list of projects",
                "description": "Listar cuentas.",
                "operationId": "show",
                "parameters": [
                    {
                        "name": "authorization",
                        "in": "header",
                        "description": "Header de autorización.",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operación correcta."
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "api_key_security_example": []
                    }
                ]
            }
        },
        "/api/cuentas/filter": {
            "get": {
                "tags": [
                    "Projects"
                ],
                "summary": "Get list of projects",
                "description": "Listar cuentas.",
                "operationId": "filter",
                "parameters": [
                    {
                        "name": "authorization",
                        "in": "header",
                        "description": "Header de autorización.",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "nombre",
                        "in": "query",
                        "description": "Nombre de la cuenta.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "estado",
                        "in": "query",
                        "description": "Estado de la cuenta.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "saldo",
                        "in": "query",
                        "description": "Saldo actual en tarjeta.",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operación correcta."
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "api_key_security_example": []
                    }
                ]
            }
        },
        "/api/cuentas/listar-estadistica": {
            "get": {
                "tags": [
                    "Projects"
                ],
                "summary": "Get list of projects",
                "description": "Listar movimientos para la estadistica.",
                "operationId": "listarEstadistica",
                "parameters": [
                    {
                        "name": "authorization",
                        "in": "header",
                        "description": "Header de autorización.",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operación correcta."
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "api_key_security_example": []
                    }
                ]
            }
        },
        "/api/movimientos/listar/{cuenta}": {
            "get": {
                "tags": [
                    "Projects"
                ],
                "summary": "Get list of projects",
                "description": "Listar movimientos.",
                "operationId": "show",
                "parameters": [
                    {
                        "name": "authorization",
                        "in": "header",
                        "description": "Header de autorización.",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "cuenta",
                        "in": "path",
                        "description": "Movimientos de esa cuenta.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operación correcta."
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "api_key_security_example": []
                    }
                ]
            }
        },
        "/api/movimientos/filter": {
            "get": {
                "tags": [
                    "Projects"
                ],
                "summary": "Get list of projects",
                "description": "Listar movimientos.",
                "operationId": "filter",
                "parameters": [
                    {
                        "name": "authorization",
                        "in": "header",
                        "description": "Header de autorización.",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "monto",
                        "in": "query",
                        "description": "Dinero transferido.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "saldo_Actual",
                        "in": "query",
                        "description": "Saldo despues de transferir.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "saldo_anterior",
                        "in": "query",
                        "description": "Saldo antes de transferir.",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operación correcta."
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "api_key_security_example": []
                    }
                ]
            }
        },
        "/api/movimientos/listar-estadistica": {
            "get": {
                "tags": [
                    "Projects"
                ],
                "summary": "Get list of projects",
                "description": "Listar movimientos para la estadistica.",
                "operationId": "listarEstadistica",
                "parameters": [
                    {
                        "name": "authorization",
                        "in": "header",
                        "description": "Header de autorización.",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operación correcta."
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "api_key_security_example": []
                    }
                ]
            }
        },
        "/api/login/{nroDoc}/{pass}": {
            "get": {
                "tags": [
                    "Projects"
                ],
                "summary": "Get list of projects",
                "description": "Iniciar sesion.",
                "operationId": "inicioSesion",
                "parameters": [
                    {
                        "name": "authorization",
                        "in": "header",
                        "description": "Header de autorización.",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "nroDoc",
                        "in": "path",
                        "description": "Nombre de usuario.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "pass",
                        "in": "path",
                        "description": "Contraseña.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operación correcta."
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "api_key_security_example": []
                    }
                ]
            }
        },
        "/api/usuarios/crear": {
            "post": {
                "tags": [
                    "Projects"
                ],
                "summary": "Get list of projects",
                "description": "Iniciar sesion.",
                "operationId": "create",
                "parameters": [
                    {
                        "name": "authorization",
                        "in": "header",
                        "description": "Header de autorización.",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "nombres",
                        "in": "query",
                        "description": "Nombre del usuario.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "apellidos",
                        "in": "query",
                        "description": "Apellidos del usuario.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "telefono",
                        "in": "query",
                        "description": "Telefono del usuario.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "correo",
                        "in": "query",
                        "description": "Correo electronico del usuario.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "usuario",
                        "in": "query",
                        "description": "Usuario.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "Contraseña del usuario.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "nro_documento",
                        "in": "query",
                        "description": "Numero de documento del usuario.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "fk_perfil",
                        "in": "query",
                        "description": "Perfil.",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": 1
                        }
                    },
                    {
                        "name": "usuario_creador",
                        "in": "query",
                        "description": "Administrador.",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operación correcta."
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "api_key_security_example": []
                    }
                ]
            }
        },
        "/api/usuarios/listar": {
            "get": {
                "tags": [
                    "Projects"
                ],
                "summary": "Get list of projects",
                "description": "Listar usuarios.",
                "operationId": "show",
                "parameters": [
                    {
                        "name": "authorization",
                        "in": "header",
                        "description": "Header de autorización.",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operación correcta."
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "api_key_security_example": []
                    }
                ]
            }
        },
        "/api/usuarios/listar/{creador}": {
            "get": {
                "tags": [
                    "Projects"
                ],
                "summary": "Get list of projects",
                "description": "Listar usuarios.",
                "operationId": "show",
                "parameters": [
                    {
                        "name": "authorization",
                        "in": "query",
                        "description": "Header de autorización.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "creador",
                        "in": "path",
                        "description": "Creador.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operación correcta."
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "api_key_security_example": []
                    }
                ]
            }
        }
    }
}